@@ -1,45 +1,54 @@
 {
   "name": "angular",
   "version": "0.0.0",
-  "private": true,
-  "dependencies": {
-    "@angular/animations": "^11.0.8",
-    "@angular/common": "^11.0.8",
-    "@angular/compiler": "^11.0.8",
-    "@angular/core": "^11.0.8",
-    "@angular/forms": "^11.0.8",
-    "@angular/platform-browser": "^11.0.8",
-    "@angular/platform-browser-dynamic": "^11.0.8",
-    "@angular/router": "^11.0.8",
-    "rxjs": "^6.6.3",
-    "tslib": "^2.1.0",
-    "zone.js": "^0.11.3"
-  },
+  "license": "MIT",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build",
+    "build": "ng build --prod",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "^11.2.0",
+    "@angular/common": "^11.2.0",
+    "@angular/compiler": "^11.2.0",
+    "@angular/core": "^11.2.0",
+    "@angular/forms": "^11.2.0",
+    "@angular/platform-browser": "^11.2.0",
+    "@angular/platform-browser-dynamic": "^11.2.0",
+    "@angular/router": "^11.2.0",
+    "core-js": "3.8.3",
+    "rxjs": "6.6.3",
+    "tslib": "2.1.0",
+    "zone.js": "0.11.3"
+  },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1100.4",
-    "@angular/cli": "~11.0.4",
-    "@angular/compiler-cli": "~11.0.4",
-    "@types/jasmine": "~3.6.0",
-    "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.1.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.0.3",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~4.0.2"
-  }
+    "@angular-devkit/build-angular": "^0.1102.0",
+    "@angular/cli": "^11.2.0",
+    "@angular/compiler-cli": "^11.2.0",
+    "@angular/language-service": "^11.2.0",
+    "@types/jasmine": "3.6.3",
+    "@types/jasminewd2": "2.0.8",
+    "@types/node": "14.14.28",
+    "codelyzer": "6.0.1",
+    "jasmine-core": "3.6.0",
+    "jasmine-spec-reporter": "6.0.0",
+    "karma": "6.1.1",
+    "karma-chrome-launcher": "3.1.0",
+    "karma-coverage-istanbul-reporter": "3.0.3",
+    "karma-jasmine": "4.0.1",
+    "karma-jasmine-html-reporter": "1.5.4",
+    "protractor": "7.0.0",
+    "ts-node": "9.1.1",
+    "tslint": "~6.1.3",
+    "typescript": "4.1.5"
+  },
+  "keywords": [
+    "angular",
+    "starter"
+  ],
+  "description": "Angular example starter project"
 }
