@@ -1,6 +1,6 @@
 import { Injectable } from "@angular/core";
-import { Http, Response } from "@angular/http";
-import { Observable } from "rxjs/Observable";
+import { HttpClient, HttpResponse } from "@angular/common/http";
+import { Observable } from "rxjs";
 
 import { Vehicle } from "./vehicle.model";
 import {
@@ -17,7 +17,7 @@ export class VehicleService {
   onDbReset = this.messageService.state;
 
   constructor(
-    private http: Http,
+    private http: HttpClient,
     private exceptionService: ExceptionService,
     private messageService: MessageService,
     private spinnerService: SpinnerService
@@ -30,44 +30,44 @@ export class VehicleService {
     this.spinnerService.show();
     return <Observable<Vehicle>>this.http
       .post(`${vehiclesUrl}`, body)
-      .map(res => <Vehicle>res.json().data)
-      .catch(this.exceptionService.catchBadResponse)
-      .finally(() => this.spinnerService.hide());
+      // .map(res => <Vehicle>res.json().data)
+      // .catch(this.exceptionService.catchBadResponse)
+      // .finally(() => this.spinnerService.hide());
   }
 
   deleteVehicle(vehicle: Vehicle) {
     this.spinnerService.show();
     return <Observable<Vehicle>>this.http
       .delete(`${vehiclesUrl}/${vehicle.id}`)
-      .map(res => this.extractData<Vehicle>(res))
-      .catch(this.exceptionService.catchBadResponse)
-      .finally(() => this.spinnerService.hide());
+      // .map(res => this.extractData<Vehicle>(res))
+      // .catch(this.exceptionService.catchBadResponse)
+      // .finally(() => this.spinnerService.hide());
   }
 
   getVehicles() {
     this.spinnerService.show();
     return <Observable<Vehicle[]>>this.http
       .get(vehiclesUrl)
-      .map(res => this.extractData<Vehicle[]>(res))
-      .catch(this.exceptionService.catchBadResponse)
-      .finally(() => this.spinnerService.hide());
+      // .map(res => this.extractData<Vehicle[]>(res))
+      // .catch(this.exceptionService.catchBadResponse)
+      // .finally(() => this.spinnerService.hide());
   }
 
-  private extractData<T>(res: Response) {
+  private extractData<T>(res: HttpResponse<any>) {
     if (res.status < 200 || res.status >= 300) {
       throw new Error("Bad response status: " + res.status);
     }
-    let body = res.json ? res.json() : null;
-    return <T>((body && body.data) || {});
+    // let body = res.json ? res.json() : null;
+    // return <T>((body && body.data) || {});
   }
 
   getVehicle(id: number) {
     this.spinnerService.show();
     return <Observable<Vehicle>>this.http
       .get(`${vehiclesUrl}/${id}`)
-      .map(res => this.extractData<Vehicle>(res))
-      .catch(this.exceptionService.catchBadResponse)
-      .finally(() => this.spinnerService.hide());
+      // .map(res => this.extractData<Vehicle>(res))
+      // .catch(this.exceptionService.catchBadResponse)
+      // .finally(() => this.spinnerService.hide());
   }
 
   updateVehicle(vehicle: Vehicle) {
@@ -76,8 +76,8 @@ export class VehicleService {
 
     return <Observable<Vehicle>>this.http
       .put(`${vehiclesUrl}/${vehicle.id}`, body)
-      .map(res => this.extractData<Vehicle>(res))
-      .catch(this.exceptionService.catchBadResponse)
-      .finally(() => this.spinnerService.hide());
+      // .map(res => this.extractData<Vehicle>(res))
+      // .catch(this.exceptionService.catchBadResponse)
+      // .finally(() => this.spinnerService.hide());
   }
 }
