@@ -1,10 +1,9 @@
 import { Component, OnDestroy, OnInit } from "@angular/core";
-import { Subscription } from "rxjs/Subscription";
+import { Subscription } from "rxjs";
 
 import { SpinnerState, SpinnerService } from "./spinner.service";
 
 @Component({
-  moduleId: module.id,
   selector: "story-spinner",
   templateUrl: "./spinner.component.html",
   styleUrls: ["./spinner.component.css"]
@@ -17,7 +16,6 @@ export class SpinnerComponent implements OnDestroy, OnInit {
   constructor(private spinnerService: SpinnerService) {}
 
   ngOnInit() {
-    componentHandler.upgradeDom();
     this.spinnerStateChanged = this.spinnerService.spinnerState.subscribe(
       (state: SpinnerState) => (this.visible = state.show)
     );
